# generated/node/src/routes/solutions.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listSolutions } from '../functions/list-solutions.function.js';
import { createSolution } from '../functions/create-solution.function.js';
import { findSolutionById } from '../functions/find-solution-by-id.function.js';
import { updateSolution } from '../functions/update-solution.function.js';
import { deleteSolution } from '../functions/delete-solution.function.js';
import { deploySolution } from '../functions/deploy-solution.function.js';
import { cloneSolutionFromTemplate } from '../functions/clone-solution-from-template.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const solutionsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/solutions', async (request, reply) => {
    try {
      const result = listSolutions(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listSolutions failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/solutions', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createSolution(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createSolution failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/solutions/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findSolutionById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findSolutionById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/solutions/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateSolution(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateSolution failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/solutions/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = deleteSolution(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'deleteSolution failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/solutions/:id/deploy', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = deploySolution(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'deploySolution failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/solutions/clone-from-template', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = cloneSolutionFromTemplate(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'cloneSolutionFromTemplate failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};