# generated/node/src/routes/traceViews.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listTraceViews } from '../functions/list-trace-views.function.js';
import { buildTraceView } from '../functions/build-trace-view.function.js';
import { findTraceViewById } from '../functions/find-trace-view-by-id.function.js';
import { updateTraceView } from '../functions/update-trace-view.function.js';
import { deleteTraceView } from '../functions/delete-trace-view.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const traceViewsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/trace-views', async (request, reply) => {
    try {
      const result = listTraceViews(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listTraceViews failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/trace-views', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = buildTraceView(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'buildTraceView failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/trace-views/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findTraceViewById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findTraceViewById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/trace-views/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateTraceView(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateTraceView failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/trace-views/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = deleteTraceView(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'deleteTraceView failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};