# generated/openapi.yaml
openapi: 3.1.0
info:
  title: Flusk API
  version: 0.1.0
paths:
...
  /api/solutions/:
    get:
      operationId: listSolutions
      tags:
        - solutions
      summary: List solutions by organization
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solution'
    post:
      operationId: createSolution
      tags:
        - solutions
      summary: Create a new solution
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solution'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Solution'
  /api/solutions/:id:
    get:
      operationId: findSolutionById
      tags:
        - solutions
      summary: Get solution by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solution'
    put:
      operationId: updateSolution
      tags:
        - solutions
      summary: Update a solution
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solution'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Solution'
    delete:
      operationId: deleteSolution
      tags:
        - solutions
      summary: Delete a solution
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solution'
  /api/solutions/:id/deploy:
    post:
      operationId: deploySolution
      tags:
        - solutions
      summary: Deploy a solution (validate, activate, version)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solution'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Solution'
  /api/solutions/clone-from-template:
    post:
      operationId: cloneSolutionFromTemplate
      tags:
        - solutions
      summary: Clone a solution from a template
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solution'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Solution'
...
  /api/trace-views/:
    get:
      operationId: listTraceViews
      tags:
        - traceViews
      summary: List trace views
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceView'
    post:
      operationId: buildTraceView
      tags:
        - traceViews
      summary: Build a trace visualization
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceView'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraceView'
  /api/trace-views/:id:
    get:
      operationId: findTraceViewById
      tags:
        - traceViews
      summary: Get trace view by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceView'
    put:
      operationId: updateTraceView
      tags:
        - traceViews
      summary: Update trace view
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceView'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraceView'
    delete:
      operationId: deleteTraceView
      tags:
        - traceViews
      summary: Delete trace view
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceView'
...
components:
  schemas:
...
    Solution:
      type: object
      properties:
        organizationId:
          type: string
          description: Organization that owns this solution
        name:
          type: string
          description: Solution name
        description:
          type: string
          description: Solution description
        type:
          $ref: '#/components/schemas/enum'
          description: Solution type
          enum:
            - agent
            - workflow
            - pipeline
            - chatbot
        status:
          $ref: '#/components/schemas/enum'
          description: Current status
          enum:
            - draft
            - active
            - paused
            - archived
        config:
          type: object
          description: Solution configuration (model, system prompt, tools, etc.)
        version:
          type: string
          description: Current version (semver)
        publishedAt:
          type: string
          description: When the solution was last published
        metrics:
          type: object
          description: Aggregated metrics (runs, cost, avg_latency)
      required:
        - organizationId
        - name
        - description
        - type
        - status
        - config
        - version
        - publishedAt
        - metrics
      description: AI solution built by users (agent, workflow, pipeline, or chatbot)
...
    TraceView:
      type: object
      properties:
        traceId:
          type: string
          description: Reference to trace entity
        viewType:
          $ref: '#/components/schemas/enum'
          description: Visualization type
          enum:
            - waterfall
            - flamegraph
            - timeline
            - tree
        totalDuration:
          type: integer
          description: Total trace duration in milliseconds
        totalCost:
          type: number
          description: Total trace cost in USD
        spanCount:
          type: integer
          description: Number of spans in the trace
        criticalPath:
          type: object
          description: List of span IDs on the critical path
        annotations:
          type: object
          description: User annotations on the trace
      required:
        - traceId
        - viewType
        - totalDuration
        - totalCost
        - spanCount
        - criticalPath
        - annotations
      description: Visualization configuration for a distributed trace
...